@@ -1,4 +1,4 @@
-class RemoveUserColumnFromCategories < ActiveRecord::Migration
+class RemoveUserColumnFromCategories < ActiveRecord::Migration[4.2]
   def up
     remove_column :categories, :user
   end
