@@ -1,4 +1,4 @@
-class CreateFriendships < ActiveRecord::Migration
+class CreateFriendships < ActiveRecord::Migration[4.2]
   def change
     create_table :friendships do |t|
       t.integer :user_id
