@@ -1,4 +1,4 @@
-class AddConfirmedToFriendships < ActiveRecord::Migration
+class AddConfirmedToFriendships < ActiveRecord::Migration[4.2]
   def change
     add_column :friendships, :confirmed, :boolean
   end
