@@ -1,4 +1,4 @@
-class AddUserToEvents < ActiveRecord::Migration
+class AddUserToEvents < ActiveRecord::Migration[4.2]
   def change
     add_column :events, :user_id, :integer
     add_index :events, :user_id
