@@ -1,4 +1,4 @@
-class AddColumnRepeatToEvents < ActiveRecord::Migration
+class AddColumnRepeatToEvents < ActiveRecord::Migration[4.2]
   def change
     add_column :events, :repeat, :string
   end
