@@ -1,4 +1,4 @@
-class AddLocationToEvents < ActiveRecord::Migration
+class AddLocationToEvents < ActiveRecord::Migration[4.2]
   def change
     add_column :events, :location, :string
   end
