@@ -1,4 +1,4 @@
-class CreateRepeatExceptionsEvents < ActiveRecord::Migration
+class CreateRepeatExceptionsEvents < ActiveRecord::Migration[4.2]
   def change
     create_table :events_repeat_exceptions, id: false do |t|
       t.belongs_to :repeat_exception, index: true
