@@ -1,4 +1,4 @@
-class AddColumnGroupIdToEventsAndCategories < ActiveRecord::Migration
+class AddColumnGroupIdToEventsAndCategories < ActiveRecord::Migration[4.2]
   def change
     add_column :events, :group_id, :integer
     add_index :events, :group_id
