@@ -1,4 +1,4 @@
-class AddColumnGroupIdToRepeatExceptions < ActiveRecord::Migration
+class AddColumnGroupIdToRepeatExceptions < ActiveRecord::Migration[4.2]
   def change
     add_column :repeat_exceptions, :group_id, :integer
     add_index :repeat_exceptions, :group_id
