@@ -1,4 +1,4 @@
-class AddPreapprovedPostsToGroups < ActiveRecord::Migration
+class AddPreapprovedPostsToGroups < ActiveRecord::Migration[4.2]
   def change
     add_column :groups, :posts_preapproved, :boolean
   end
