@@ -1,4 +1,4 @@
-class DropFriends < ActiveRecord::Migration
+class DropFriends < ActiveRecord::Migration[4.2]
   def change
     drop_table :friendships
   end
