@@ -1,4 +1,4 @@
-class AddAvatarToUsers < ActiveRecord::Migration
+class AddAvatarToUsers < ActiveRecord::Migration[4.2]
 	def self.up
 	  add_attachment :users, :avatar
 	end
