@@ -1,4 +1,4 @@
-class AddAvatarToGroups < ActiveRecord::Migration
+class AddAvatarToGroups < ActiveRecord::Migration[4.2]
   def change
     add_column :groups, :avatar_file_name, :string
     add_column :groups, :avatar_content_type, :string
