@@ -1,6 +1,6 @@
 class UsersController < ApplicationController
-  before_action :authorize_admin, only: [:index]
-  before_action :set_user, only: [:show]
+  before_action :authorize_admin!, only: [:index, :promote, :demote, :inspect]
+  before_action :authorize_signed_in!, only: [:destroy]
 
   def index
     @users = User.all.sort_by(&:name)
@@ -8,8 +8,17 @@ class UsersController < ApplicationController
   end
 
   def show
+
+    @user = User.from_param(params[:id])
+
     # is this a user looking at their own profile?
     @profile = current_user == @user
+
+    # forces custom urls to be displayed (when applicable)
+    if params[:id].is_int? and @user.has_custom_url?
+      redirect_to user_path(@user), status: :moved_permanently
+    end
+
 
     @current_page = params[:page]&.to_sym || :schedule
     case @current_page
@@ -35,30 +44,54 @@ class UsersController < ApplicationController
     q = params[:q]&.strip
     return unless q.present?
 
-    if request.path_parameters[:format] == 'json'
-      @users = User.where('name LIKE ?', "%#{q}%").limit(10)
-      @users = User.rank(@users, q)
-    else
-      @users = User.where('name LIKE ?', "%#{q}%")
+    if q != "" # only search if it's not silly
+      if request.path_parameters[:format] == 'json'
+        @users = User.where('name LIKE ?', "%#{q}%").limit(10)
+        @users = User.rank(@users, q)
+      else
+        @users = User.where('name LIKE ?', "%#{q}%")
+      end
     end
 
     respond_to do |format|
       format.html
       format.json do
+        # Return the users in their public JSON form
         user_map = @users.map(&:convert_to_json)
+
         render json: user_map
       end
     end
   end
 
-  private
+  def promote
+    @user = User.find(params[:id])
+    @user.update(admin: true)
+    render json: { action: "promote", uid: @user.id, new_href: demote_user_path(@user)}
+  end
 
-  def set_user
-    if params[:id_or_url] =~ User.REGEX_USER_ID
-      @user = User.find_by! id: params[:id_or_url]
-      redirect_to user_path(@user) if @user.has_custom_url?
+  def demote
+    @user = User.find(params[:id])
+    @user.update(admin: false)
+    render json: { action: "demote", uid: @user.id, new_href: promote_user_path(@user)}
+  end
+
+  def destroy
+    @user = User.from_param(params[:id])
+
+    if current_user.admin
+      @user.destroy
+      redirect_to admin_panel_path, notice: "User deleted"
+    elsif current_user == @user
+      @user.destroy
+      redirect_to home_path, notice: "Account deleted"
     else
-      @user = User.find_by! custom_url: params[:id_or_url]
+      redirect_to user_session_path, alert: "You don't have permission to do that!"
     end
+  end
+
+  def inspect
+    @user = User.find(params[:id])
+    @user_groups = UsersGroup.where(user_id: @user.id)
   end
 end
